Refactored add or edit user component into one
